% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adipose.R
\docType{data}
\name{adipose}
\alias{adipose}
\title{BMI and Visceral Adipose Tissue Mass}
\format{
A data frame with 37 observations of 7 variables.
\describe{
  \item{sex}{Male or Female}
  \item{health}{Description of general health of participant}
  \item{age}{age in years}
  \item{ldl}{LDL cholesterol in mmol/L}
  \item{hdl}{HDL cholesterol in mmol/L}
  \item{trig}{Triglycerides in mmol/L}
  \item{glucose}{Blood glucose in mmol/L}
  \item{stature_cm}{height in centimeters}
  \item{waist_cm}{waist measurement in cm}
  \item{hips_cm}{hips measurement in cm}
  \item{vis_fat_mass}{visceral adipose fat mass in grams}
  \item{bmi}{body mass index kg/m^2}
}
}
\source{
Swainson MG, Batterham AM, Tsakirides C, Rutherford ZH, Hind K (2017) Prediction of whole-body fat percentage and visceral adipose tissue mass from five anthropometric variables. PLoS ONE 12(5): e0177175. https://doi.org/10.1371/journal.pone.0177175
}
\usage{
adipose
}
\description{
Measurements of BMI and adipose tissue for adults.
}
\details{
From the authors: "The aims of this study were (a) to compare five anthropometric predictors of %FM and VAT mass, and (b) to explore new cut-points for the best of these predictors to improve the characterisation of obesity."
}
\keyword{datasets}
