% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cows.R
\docType{data}
\name{cows_small}
\alias{cows_small}
\title{Cooling down hot cows}
\format{
A data frame with 798 observations of 9 variables
\describe{
  \item{cow}{ID of the cow that was measured.}
  \item{control}{Amount that the cow's shoulder cooled down (degrees Celsius) in 3 minutes when no water was sprayed.}
  \item{tk_0_75}{Cool down in degrees C after 3 minutes using nozzle TK-0.75}
  \item{tk_12}{Cool down in degrees C after 3 minutes using nozzle TK-12}
}
}
\source{
Tucker, Cassandra B., Schütz, Karin E., & Van Os (Chen), Jennifer M. (2020). Data from: Cooling cows efficiently with sprinklers: physiological responses to water spray [Data set].

\url{https://zenodo.org/record/3862181#.XwTRNpNKjEa}
}
\usage{
cows_small
}
\description{
Cows in California were sprayed down with water using nozzle type TK-0.75. The temperature of the cows' shoulders were
measured after 3 minutes.
}
\details{
From the authors: "Dairies in the United States commonly cool cattle with sprinklers mounted over the feed bunk that intermittently spray the cows' backs. These systems use potable water—an increasingly scarce resource—but there is little experimental evidence about how much is needed to cool cows or about droplet size, which is thought to affect hair coat penetration. "
}
\keyword{datasets}
