% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_period.R
\docType{data}
\name{crit_period}
\alias{crit_period}
\title{Critical Period Hypothesis}
\format{
A data frame with 30 observations of 2 variables. Descriptions of variables are from the authors.
\describe{
  \item{aoa}{age of onset of acquisition of second language. In years.}
  \item{gjt}{grammaticality judgement test. A measure of second language proficiency.}
  \item{locale}{Factor with two levels. "North America" and "Israel"}
}
}
\source{
Vanhove J (2013) The Critical Period Hypothesis in Second Language Acquisition: A Statistical Critique and a Reanalysis. PLoS ONE 8(7): e69172.

\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0069172}
}
\usage{
crit_period
}
\description{
The age of onset of acquisition and ultimate level of attainment of second language learners in North America and Israel.
}
\details{
From the author: "In second language acquisition research, the critical period hypothesis (cph) holds that the function between learners' age and their susceptibility to second language input is non-linear. This paper revisits the indistinctness found in the literature with regard to this hypothesis's scope and predictions."
}
\keyword{datasets}
