% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\docType{data}
\name{draft}
\alias{draft}
\title{The 1970 Draft Lottery}
\format{
A data frame with 366 observations of 5 variables.
\describe{
  \item{Day}{Day of the month.}
  \item{Month}{Month abbreviation.}
  \item{MonthNo}{Month number 1-12.}
  \item{DayofYear}{Day of year from 1-366.}
  \item{DraftNo}{Draft number from 1-366. Lowest numbers were drafted first.}
}
}
\source{
United States Government Selective Service System.  https://www.sss.gov/history-and-records/vietnam-lotteries/
}
\usage{
draft
}
\description{
Results of the draft lottery held December 1, 1969 to determine the order in which young men would be drafted in 1970 for service in the Vietnam war.
}
\keyword{datasets}
