% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_dogs.R
\docType{data}
\name{hot_dogs}
\alias{hot_dogs}
\title{Hot Dog Nutrition}
\format{A data frame with 54 observations of 3 variables.
\describe{
  \item{type}{Factor with levels Beef, Meat or Poultry}
  \item{calories}{Number of calories in the hot dog (kC)}
  \item{sodium}{Sodium content of the hot dog (mg)}
}}
\source{
Moore, David S., and George P. McCabe (1989). Introduction to the Practice of Statistics. Original source: Consumer Reports, June 1986, pp. 366-367.

\url{http://wiki.stat.ucla.edu/socr/index.php/SOCR_012708_ID_Data_HotDogs}
}
\usage{
hot_dogs
}
\description{
Calorie and sodium content of hot dogs of various types of meats.
}
\details{
From the web page: "Results of a laboratory analysis of calories and sodium content of major hot dog brands. Researchers for Consumer Reports analyzed three types of hot dog: beef, poultry, and meat (mostly pork and beef, but up to 15% poultry meat)."
}
\keyword{datasets}
