% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_pot.R
\docType{data}
\name{mice_pot}
\alias{mice_pot}
\title{Mice on pot}
\format{
A data frame with 47 observations of 2 variables
\describe{
  \item{group}{One of five levels, VEH (control) or the amount of THC in mg/kg the mouse was injected with.}
  \item{percent_of_act}{Percent of baseline locomotion activity}
}
}
\source{
Viñals X, Moreno E, Lanfumey L, Cordomí A, Pastor A, de La Torre R, et al. (2015) Cognitive Impairment Induced by Delta9-tetrahydrocannabinol Occurs through Heteromers between Cannabinoid CB1 and Serotonin 5-HT2A Receptors. PLoS Biol 13(7): e1002194. https://doi.org/10.1371/journal.pbio.1002194
}
\usage{
mice_pot
}
\description{
Wild type mice were injected with THC at various doses based on the weight of the mouse. The percent of baseline locomotion
was observed for each mouse.
}
\details{
There is much more data of this type available in the paper. The authors tested the mice in numerous ways,
and they also compared wild type mice with serotonin 2A receptors knockout mice.

From the authors: "Activation of cannabinoid CB1 receptors (CB1R) by delta9-tetrahydrocannabinol (THC) produces a variety of negative effects with major consequences in cannabis users that constitute important drawbacks for the use of cannabinoids as therapeutic agents. For this reason, there is a tremendous medical interest in harnessing the beneficial effects of THC. Behavioral studies carried out in mice lacking 5-HT2A receptors (5-HT2AR) revealed a remarkable 5-HT2AR-dependent dissociation in the beneficial antinociceptive effects of THC and its detrimental amnesic properties."
}
\keyword{datasets}
