% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerball.R
\docType{data}
\name{powerball}
\alias{powerball}
\title{Powerball draws from 2010-2020}
\format{
A data frame with 1094 observations of 8 variables.
\describe{
  \item{Draw.Date}{Date of the drawing.}
  \item{Ball1-Ball6}{The numbers on the six balls drawn. Ball6 is the powerball.}
  \item{Multiplier}{A randomly drawn value that will multiply a player's winnings, should they pay extra for a power play ticket.}
}
}
\source{
data.gov, https://catalog.data.gov/dataset/lottery-powerball-winning-numbers-beginning-2010
}
\usage{
powerball
}
\description{
The results of all Powerball lottery draws from February 3, 2010 to July 29, 2020.
}
\details{
Powerball is an American lottery game offered in most U.S. states.
Drawings happen twice weekly, and consist of five white balls drawn without replacement and one red ball, called the Powerball.
The format of ball drawings changed on January 15, 2012 and again on October 4, 2015.
}
\keyword{datasets}
