% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotland_births.R
\docType{data}
\name{scotland_births}
\alias{scotland_births}
\title{Births in Scotland}
\format{
A data frame with 45 observations of 76 variables.
\describe{
  \item{age}{Age of the mother, in years}
  \item{x****}{Year of birth of baby; 75 columns of this type from 1945-2019.}
}
}
\source{
National Records of Scotland, https://www.nrscotland.gov.uk/statistics-and-data/statistics/statistics-by-theme/vital-events/births/births-time-series-data
}
\usage{
scotland_births
}
\description{
The number of babies born in Scotland from 1945-2019, by age of the mother.
}
\keyword{datasets}
