% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble.R
\docType{data}
\name{scrabble}
\alias{scrabble}
\title{Scrabble tile values}
\format{
A data frame with 100 observations of 2 variables
\describe{
  \item{piece}{the letter A-Z shown on the tile or "blank"}
  \item{points}{the points shown on the tile}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Scrabble_letter_distributions}
}
\usage{
scrabble
}
\description{
This dataset contains 100 records corresponding to the 100 tiles in English language editions of the board game Scrabble.
}
\keyword{datasets}
