% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble.R
\docType{data}
\name{scrabble_ml}
\alias{scrabble_ml}
\title{Scrabble tile values}
\format{
A data frame with 613 observations of 3 variables
\describe{
  \item{lang}{language of the scrabble tiles; one of dutch, english, french, l33t, malagasy, or tswana}
  \item{piece}{the letter A-Z (or number in the case of l33t) shown on the tile or "blank." Diacritics are ignored.}
  \item{points}{the points shown on the tile}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Scrabble_letter_distributions}
}
\usage{
scrabble_ml
}
\description{
This dataset contains 613 observations corresponding to the tiles in editions of the board game Scrabble for 6 languages.
}
\keyword{datasets}
