% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seoulweather.R
\docType{data}
\name{seoulweather}
\alias{seoulweather}
\title{Weather data for Seoul}
\format{
A data frame with 7752 observations of 25 variables
\describe{
}
}
\source{
'Cho, D., Yoo, C., Im, J., & Cha, D. (2020). Comparative assessment of various machine learning-based bias correction methods for numerical weather prediction model forecasts of extreme air temperatures in urban areas. Earth and Space Science.'

\url{https://archive.ics.uci.edu/ml/datasets/Bias+correction+of+numerical+prediction+model+temperature+forecast}
}
\usage{
seoulweather
}
\description{
The forecast for the next day and current day weather conditions in and near Seoul from 2013-06-30 to 2017-08-30.
}
\details{
From the source: "This data is for the purpose of bias correction of next-day maximum and minimum air temperatures forecast of the LDAPS model operated by the Korea Meteorological Administration over Seoul, South Korea. This data consists of summer data from 2013 to 2017. The input data is largely composed of the LDAPS model's next-day forecast data, in-situ maximum and minimum temperatures of present-day, and geographic auxiliary variables. There are two outputs (i.e. next-day maximum and minimum air temperatures) in this data."
}
\keyword{datasets}
