% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seoulweather.R
\docType{data}
\name{seoulweather}
\alias{seoulweather}
\title{Weather data for Seoul}
\format{
A data frame with 7752 observations of 25 variables
\describe{
}
}
\source{
'Cho, D., Yoo, C., Im, J., & Cha, D. (2020). Comparative assessment of various machine learning-based bias correction methods for numerical weather prediction model forecasts of extreme air temperatures in urban areas. Earth and Space Science.'

\url{https://archive.ics.uci.edu/ml/datasets/Bias+correction+of+numerical+prediction+model+temperature+forecast}
}
\usage{
seoulweather
}
\description{
The forecast for the next day and current day weather conditions in and near Seoul from 2013-06-30 to 2017-08-30.
}
\keyword{datasets}
